requirements.txt
flask
requests
asyncio
aiohttp
pymysql
SQLAlchemy
python-dotenv